+++ corpus/existing_aliases/as_alias_example.clj
@@ -1,0 +1,4 @@
+(ns as-alias-example
+  (:require [baz.qux :as-alias q]))
+
+(identity ::baz.qux/some-fn)